# public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sample API",
        "description": "An example API for demonstration purposes",
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "schemes": ["http"],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Create a new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Login",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Get a list of users",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authentication",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a user",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "Users array",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "users": {
                                    "type": "array"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User updated successfully"
                    }
                }
            }
        },
        "/users/{id}/delete": {
            "delete": {
                "summary": "Delete a user",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token for authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "A user ID",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    }
                }
            }
        }
    }
}